Moved library class
